bug/`qs pr` on Jira issue does not work
