+++ internal/jira/const.go
@@ -1,0 +1,5 @@
+package jira
+
+const (
+	jiraDomain = "https://untill.atlassian.net"
+)
